improve the readablity of the Main State
